Validation for website

+++ public/assignment/views/website/website-edit.view.controller.js
@@ -3,7 +3,7 @@
         .module('WebAppMaker')
         .controller('EditWebsiteController', EditWebsiteController);
 
-    function EditWebsiteController($location, $routeParams, WebsiteService) {
+    function EditWebsiteController($location, $routeParams, $window, WebsiteService) {
         var vm = this;
         vm.userId = $routeParams.userId;
         vm.websiteId = $routeParams.websiteId;
@@ -22,23 +22,32 @@
         init();
 
         function updateWebsite() {
-            WebsiteService.updateWebsite(vm.websiteId, vm.userId, vm.website)
-                .then(function (website) {
-                        $location.url('/user/' + vm.userId + '/website');
-                    },
-                    function (err) {
-                        vm.msg = {type: 'error', text: err.body};
-                    });
+            if (!vm.website.name) {
+                vm.msg = {type: 'error', text: 'Website name required'};
+            }
+            else {
+                WebsiteService.updateWebsite(vm.websiteId, vm.userId, vm.website)
+                    .then(function (website) {
+                            $location.url('/user/' + vm.userId + '/website');
+                        },
+                        function (err) {
+                            vm.msg = {type: 'error', text: err.body};
+                        });
+            }
         }
 
         function deleteWebsite(websiteId) {
-            WebsiteService.deleteWebsite(websiteId, vm.userId)
-                .then(function () {
-                    $location.url('/user/' + vm.userId + '/website');
+            var choice = $window.confirm('Are you sure you want to delete?');
+            if (choice) {
+                WebsiteService.deleteWebsite(websiteId, vm.userId)
+                    .then(function () {
+                        $location.url('/user/' + vm.userId + '/website');
 
-                }, function (err) {
-                    vm.msg = {type: 'error', text: err.body};
-                });
+                    }, function (err) {
+                        vm.msg = {type: 'error', text: err.body};
+                    });
+            }
+
         }
     }
 })();
